refactor: move CreateLink to separate component
